second commit

+++ Banking1/src/main/java/com/lti/banking/dao/CustomerDAOImpl.java
@@ -16,7 +16,7 @@ public class CustomerDAOImpl implements CustomerDAO{
 	
 	@Override
 	@Transactional
-	public int addCustomer(Customer cus) {
+	public long addCustomer(Customer cus) {
 		
          em.persist(cus);
