+++ Banking1/src/main/java/com/lti/banking/service/CustomerService.java
@@ -4,6 +4,6 @@ import com.lti.banking.beans.Customer;
 
 public interface CustomerService {
 
-	public int addCustomer(Customer cus);
+	public long addCustomer(Customer cus);
 	public Customer searchById(int id);
 }
